Split NDI's producer/consumer code
